Add support for workspace.dependencies.
Rust 1.64 added workspace inheritance, which allowed dependencies in
crates to be expressed with 'foo.workspace = true' syntax.

The toml parser in the crates extension would misinterpret these lines
and throw an error message next to them.

With this patch, the parser skips dependencies that contain the word
"workspace", only parsing these dependencies in the workspace global
toml file that defines the workspace-level dependency.

This resolves #168.